backend/requirements.txt
fastapi
uvicorn
pymongo
pandas
python-multipart
google-generativeai
python-dotenv
scikit-learn
numpy
reportlab
python-jose
passlib
